Move return type backend from types to query folder
