+++ src/stories/Dashboard.stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Meta } from '@storybook/react/types-6-0';
-import { Story } from '@storybook/react';
 import Dashboard from '../components/organisms/Dashboard';
 
 export const Primary: React.FC = () => <Dashboard />;
