+++ src/stories/ThresholdChart.stories.tsx
@@ -3,7 +3,7 @@ import { Meta } from '@storybook/react/types-6-0';
 import ThresholdChart, { ThresholdChartProps } from '../components/atoms/ThresholdChart';
 import { Story } from '@storybook/react';
 import { MetApiCompactForecast } from '../mockdata/metApi';
-import { MetApiCompactAirTemperature } from '../types/compact';
+import { MetApiCompactAirTemperature } from '../queries/metApi';
 import { appleStock } from '@visx/mock-data';
 
 export default {
