--- src/types/compact.ts
@@ -1,19 +1,0 @@
-export type MetApiCompactAirTemperature = {
-    forecast: {
-        forecastProperties: {
-            meta: {
-                updatedAt: string;
-            };
-            timeseries: {
-                time: string;
-                forecastData: {
-                    instant: {
-                        details: {
-                            airTemperature: number;
-                        };
-                    };
-                };
-            }[];
-        };
-    };
-};
